change axis attribute from int to int64_t

+++ src/targets/gpu/device/argmax.cpp
@@ -12,7 +12,7 @@ inline namespace MIGRAPHX_INLINE_NS {
 namespace gpu {
 namespace device {
 
-void argmax(hipStream_t stream, const argument& result, const argument& arg, int axis)
+void argmax(hipStream_t stream, const argument& result, const argument& arg, int64_t axis)
 {
     arg_op(argmax_op{}, stream, result, arg, axis);
 }
